refactor(test): split scroll-edge-state suite, drop signature test-gaming
Splits the 1,323-line scroll-edge-state.test.ts into three behavior-focused
sibling suites (container selection, scope naming, pass orchestration) plus a
shared fixtures module, per AGENTS.md's per-file LOC tripwire and sibling-test
convention.

Also removes the dead `signature`/buildScrollStateSignature/collectSubtreeNodes/
hasAncestor production code: it had zero consumers anywhere in the repo, and
the tests pinning its private string format were the mutation-score-gaming
flagged in review — with it gone, there's nothing left to game.

Addresses review feedback on PR #1474 (#1455).
